feat(us_2005): implementation of a functionality to list grades (fixes #72)
In order to implement this functionality, I created an Entity called
ExamResult along with its respective value objects, previously discussed
and designed in the domain model. Furthermore, I tested all the new
classes and developed a service called ListExamResultsService for which
I developed a method to be used in us_2006. Finally, I bootstrapped
ExamResult information for a student with the username mary.